requirements.txt
openai==0.26.1
PyAudio==0.2.13
python-dotenv==0.21.1
pyttsx3==2.90
pywin32==305
SpeechRecognition==3.9.0